requirements.txt
django
djangorestframework
pytest-django
gunicorn
pillow
whitenoise
django-cors-headers